Rename; report analytics for internal page nav
