+++ src/components/MultiPageTutorials/MPTNextLinkMini.js
@@ -6,8 +6,8 @@ import { theme } from '../../theme/docsTheme';
 import Link from '../Link';
 import { navLinkButtonStyle } from '../InternalPageNav';
 import { LINK_TITLE } from './constants';
-import { useShouldShowNext } from './hooks/useShouldShowNext';
-import { useActiveMpTutorial } from './hooks/useActiveMpTutorial';
+import { useShouldShowNext } from './hooks/use-should-show-next';
+import { useActiveMpTutorial } from './hooks/use-active-mp-tutorial';
 
 const baseStyle = css`
   float: right;
