+++ src/components/MultiPageTutorials/index.js
@@ -2,4 +2,4 @@ export { MPTNextLinkFull } from './MPTNextLinkFull';
 export { MPTNextLinkMini } from './MPTNextLinkMini';
 export { StepNumber } from './StepNumber';
 export { TimeRequired } from './TimeRequired';
-export { useActiveMpTutorial } from './hooks/useActiveMpTutorial';
+export { useActiveMpTutorial } from './hooks/use-active-mp-tutorial';
